pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-live-bot"
version = "0.1.0"
description = "Safety-first Polymarket CLOB live trading bot."
requires-python = ">=3.11"
dependencies = [
  "aiohttp==3.10.11",
  "alembic==1.13.3",
  "asyncpg==0.30.0",
  "fastapi==0.115.6",
  "httpx==0.27.2",
  "prometheus-client==0.21.1",
  "py-clob-client==0.23.0",
  "pydantic==2.10.4",
  "pydantic-settings==2.7.0",
  "python-json-logger==2.0.7",
  "PyYAML==6.0.2",
  "SQLAlchemy==2.0.36",
  "tenacity==9.0.0",
  "uvicorn[standard]==0.34.0",
  "websockets==14.1"
]

[project.optional-dependencies]
dev = [
  "mypy==1.13.0",
  "pytest==8.3.4",
  "pytest-asyncio==0.25.0",
  "respx==0.21.1",
  "ruff==0.8.4"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
